Skip serializing estimators + fix test + added empty data transform test
